fix(cast): improve cast wallet new
